server/pyproject.toml
[project]
name = "doctor-management"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Joshua", email = "josuaprakasam04@gmail.com"},
]
dependencies = [
    "fastapi>=0.111.1",
    "sqlalchemy>=2.0.31",
    "pydantic-settings>=2.3.4",
    "psycopg[binary,pool]>=3.2.1",
    "pyjwt>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
